Moved SocketVision to /robot/util
